requirements.txt
# pip install -r requirements.txt

# ------------------------------------------
# 🌐 Django и сопутствующие пакеты
# ------------------------------------------
Django>=5.0.0

# ------------------------------------------
# 📊 Обработка данных
# ------------------------------------------
pandas>=2.0.0              # Построение DataFrame из EntryValue и чтение Excel
numpy>=1.26.0              # Поддержка числовых операций, требуется pandas

# ------------------------------------------
# 🤖 Машинное обучение (линейная модель)
# ------------------------------------------
scikit-learn>=1.4.0        # Используется для LinearRegression (базовая модель)
joblib>=1.2              # Сохраняем и загружаем модели в формате .pkl

# ------------------------------------------
# 📥 Загрузка Excel-файлов через админку
# ------------------------------------------
openpyxl>=3.1            # Движок для чтения .xlsx в pandas.read_excel()
python-slugify>=8.0.0  # 🔤 Преобразует строку (например, "Тошнота сильная!") в slug (например, "toshnota-silnaya")
                     # Используется для генерации уникального ключа `key`
gunicorn>=20.1.0 # ⚠️ Не забудь убедиться, что gunicorn установлен в requirements.txt.